+++ api/server/runner_httptrigger_test.go
@@ -383,7 +383,7 @@ func TestTriggerRunnerTimeout(t *testing.T) {
 	}()
 
 	models.MaxMemory = uint64(1024 * 1024 * 1024) // 1024 TB
-	hugeMem := uint64(models.MaxMemory - 1)
+	hugeMem := models.MaxMemory - 1
 
 	app := &models.App{ID: "app_id", Name: "myapp", Config: models.Config{}}
 	fn := &models.Fn{ID: "hot", Name: "hot", AppID: app.ID, Image: "fnproject/fn-test-utils", ResourceConfig: models.ResourceConfig{Memory: 128, Timeout: 4, IdleTimeout: 30}}
